Relax some production requirements
If we use the production environment whenever we run in a container,
these changes relax some of the constraints so it can be run locally on
developer machines. The downside is that this removes some of the safety
checks that we're not using insecure defaults when pushing to a live
environment.