Use patch Rails to 6.0.3.7 and use workaround for actionpack to mitigate CVEs [MENU-1658] (#180)
* Patch rails to 6.0.3.7 from 6.0.3.5
* Add workaround initializer for actionpack CVE-2021-22942

Fully updating actionpack itself would require updating rails from 6.0.3.5 to 6.0.4.1, and even though it's a patch update, that could get hairy. I've opted for now to upgrade only to 6.0.3.7 (patch of a patch level update), which resolves CVE-2021-22904, CVE-2021-22885, and CVE-2021-22902. I'm mitigating the final CVE using the initializer provided as a workaround with the resolution of CVE-2021-22942